Rename QueriesService to DataService. Add services folder
