+++ code/tools/WtsTelemetry/WtsTelemetry/Queries.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace WtsTelemetry
+﻿namespace WtsTelemetry
 {
     public static class Queries
     {
